fix Login, Sign up, Add LiveStream Screen(Optional) and feeling free
Add kiểm tra điều kiện input login,sign up
fix lỗi nhập input auto bị xóa của form login
Add LiveStream Screen...
Có add thêm thư viện nên khi chạy nhớ npm i nhé ;)
