+++ Utils/Connection/AppInfo.js
@@ -1,6 +1,6 @@
 const AppID = '234b6f411ca94063b20c46252c2843ca';
 const Default_Channel_Name ='Chung';
-const Token_Daily= '006234b6f411ca94063b20c46252c2843caIACKwa8awfDqeeCBicXOwk4LKSLPR+v1R/TpY5sDZsBRYQEScV0AAAAAEADEGoyZ0wtzYQEAAQB8C3Nh';
+const Token_Daily= '006234b6f411ca94063b20c46252c2843caIABQjaJpnb5SjfaDbVnO42d+gHIFpDsu5KRA3XEGGbtyQQEScV0AAAAAEACHtvL/qZB3YQEAAQBJkHdh';
 
 module.exports = {
     AppID,
