+++ Utils/Connection/SongSelector.js
@@ -1,0 +1,27 @@
+import React, { useState } from "react";
+import { Button, Text, View } from "react-native";
+import Modal from "react-native-modal";
+
+function SongSelector() {
+  const [isModalVisible, setModalVisible] = useState(false);
+
+  const toggleModal = () => {
+    setModalVisible(!isModalVisible);
+  };
+
+  return (
+    <View style={{ flex: 1 }}>
+      <Button title="Show modal" onPress={toggleModal} />
+
+      <Modal isVisible={isModalVisible}>
+        <View style={{ flex: 1 }}>
+          <Text>Hello!</Text>
+
+          <Button title="Hide modal" onPress={toggleModal} />
+        </View>
+      </Modal>
+    </View>
+  );
+}
+
+export default SongSelector;
