+++ package.json
@@ -23,11 +23,13 @@
     "react-native": "0.66.0",
     "react-native-agora": "^3.5.0",
     "react-native-gesture-handler": "^1.10.3",
+    "react-native-modal": "^13.0.0",
     "react-native-nodemediaclient": "^0.2.17",
     "react-native-safe-area-context": "^3.3.2",
     "react-native-screens": "^3.8.0",
     "react-native-swiper": "^1.6.0",
-    "react-native-web": "^0.17.5"
+    "react-native-web": "^0.17.5",
+    "react-navigation": "^4.4.4"
   },
   "devDependencies": {
     "@babel/core": "^7.15.8",
